+++ NEWS.md
@@ -7,11 +7,19 @@
 * `multi_join()`: When joining multiple data frames on different variable names it is now possible to pass in a list of vectors for the first `on` list entry to enable joining each data frame on the first one on different variable names. (21.09.2026)
 * `drop_type_vars()`: When passing in a list of data frames, the function now removes the variables `TYPE`, `TYPE_NR` and `DEPTH` from all data frames within the list. (22.09.2026)
 * `retain_stat()`: Can now generate cumulative values. (23.09.2026)
+* `multi_join()`: Single variables passed to the `on` parameter can now be written without quotation marks. (25.09.2026)
+* `any_table()`, `excel_output_style()`: The new `block_borders` style parameter enables to draw borders between statistic blocks within the table area. (25.09.2026)
 
 ### Fixed
 
 * `compute.()`: Some custom functions didn't work consistently in different situations. This is fixed now. (21.09.2026)
 * Most of the functions couldn't resolve custom parameters passed on from a wrapper function by themselves. This is hopefully fixed. (22.09.2026)
+* `compute.()`: Doesn't error anymore when calculations are written on multiple lines. (24.09.2026)
+* `any_table()`: Variable labels weren't set when the variable name contained a `statistics` keyword. This is fixed now. (25.09.2026)
+* `any_table()`: When working with pre summarised data then now the value variables appear in user provided order instead of order of appearance within the given data frame. (25.09.2026)
+* `running_number()`: Now considers all variables when passing a vector of variable names into `by`. (25.09.2026)
+* `if.()`, `else_if.()`, `ifelse_multi()`, `where.()`: Don't error anymore on parsing variable names ending in a number. (25.09.2026)
+* `summarise_plus()`: In case the function was used inside a custom function, percentiles weren't detected as `statistics`. This is fixed now. (25.09.2026)
 
 # qol 1.3.5